Separate Upgrade/Version-related code from ConfigManager. (#3497)
