+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/infrastructure/CorfuReplicationDiscoveryService.java
@@ -22,6 +22,7 @@ import org.corfudb.infrastructure.logreplication.proto.LogReplicationMetadata.Re
 import org.corfudb.infrastructure.logreplication.proto.LogReplicationMetadata.ReplicationEventKey;
 import org.corfudb.infrastructure.logreplication.replication.receive.LogReplicationMetadataManager;
 import org.corfudb.infrastructure.logreplication.utils.LogReplicationConfigManager;
+import org.corfudb.infrastructure.logreplication.utils.LogReplicationUpgradeManager;
 import org.corfudb.runtime.CorfuRuntime;
 import org.corfudb.runtime.exceptions.RetryExhaustedException;
 import org.corfudb.runtime.exceptions.TransactionAbortedException;
@@ -108,6 +109,11 @@ public class CorfuReplicationDiscoveryService implements CorfuReplicationDiscove
      * in Log Replication
      */
     private LogReplicationConfigManager replicationConfigManager;
+
+    /**
+     * Responsible for version management
+     */
+    private LogReplicationUpgradeManager upgradeManager;
 
     /**
      * Used by the source cluster to initiate Log Replication
@@ -353,6 +359,7 @@ public class CorfuReplicationDiscoveryService implements CorfuReplicationDiscove
         // Through the config manager, retrieve system-specific configurations such as streams to replicate
         // for supported replication models and version
         replicationConfigManager = new LogReplicationConfigManager(getCorfuRuntime(), serverContext);
+        upgradeManager = new LogReplicationUpgradeManager(getCorfuRuntime(), serverContext.getPluginConfigFilePath());
         logReplicationConfig = replicationConfigManager.getConfig();
 
         Set<String> remoteClusterIds = new HashSet<>();
@@ -502,7 +509,7 @@ public class CorfuReplicationDiscoveryService implements CorfuReplicationDiscove
                 if (replicationManager == null) {
                     replicationManager = new CorfuReplicationManager(replicationContext, localNodeDescriptor,
                         remoteSessionToMetadataManagerMap, serverContext.getPluginConfigFilePath(), getCorfuRuntime(),
-                        replicationConfigManager);
+                        replicationConfigManager, upgradeManager);
                 } else {
                     // Replication Context contains the topology which
                     // must be updated if it has changed
