+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/infrastructure/CorfuReplicationManager.java
@@ -6,9 +6,8 @@ import org.corfudb.infrastructure.LogReplicationRuntimeParameters;
 import org.corfudb.infrastructure.logreplication.replication.receive.LogReplicationMetadataManager;
 import org.corfudb.infrastructure.logreplication.runtime.CorfuLogReplicationRuntime;
 import org.corfudb.infrastructure.logreplication.utils.LogReplicationConfigManager;
+import org.corfudb.infrastructure.logreplication.utils.LogReplicationUpgradeManager;
 import org.corfudb.runtime.CorfuRuntime;
-import org.corfudb.runtime.exceptions.TransactionAbortedException;
-import org.corfudb.runtime.exceptions.unrecoverable.UnrecoverableCorfuInterruptedError;
 import org.corfudb.util.retry.IRetry;
 import org.corfudb.util.retry.IntervalRetry;
 import org.corfudb.util.retry.RetryNeededException;
@@ -39,19 +38,23 @@ public class CorfuReplicationManager {
 
     private final LogReplicationConfigManager replicationConfigManager;
 
+    private final LogReplicationUpgradeManager upgradeManager;
+
     /**
      * Constructor
      */
     public CorfuReplicationManager(LogReplicationContext context, NodeDescriptor localNodeDescriptor,
                                    Map<ReplicationSession, LogReplicationMetadataManager> metadataManagerMap,
                                    String pluginFilePath, CorfuRuntime corfuRuntime,
-                                   LogReplicationConfigManager replicationConfigManager) {
+                                   LogReplicationConfigManager replicationConfigManager,
+                                   LogReplicationUpgradeManager upgradeManager) {
         this.context = context;
         this.metadataManagerMap = metadataManagerMap;
         this.pluginFilePath = pluginFilePath;
         this.corfuRuntime = corfuRuntime;
         this.localNodeDescriptor = localNodeDescriptor;
         this.replicationConfigManager = replicationConfigManager;
+        this.upgradeManager = upgradeManager;
     }
 
     /**
@@ -131,7 +134,7 @@ public class CorfuReplicationManager {
                             .maxWriteSize(corfuRuntime.getParameters().getMaxWriteSize())
                             .build();
                     CorfuLogReplicationRuntime replicationRuntime = new CorfuLogReplicationRuntime(parameters,
-                        metadataManagerMap.get(session), replicationConfigManager, session);
+                        metadataManagerMap.get(session), upgradeManager, replicationConfigManager, session);
                     replicationRuntime.start();
                     runtimeToRemoteSession.put(session, replicationRuntime);
                 } catch (Exception e) {
