+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/replication/LogReplicationSourceManager.java
@@ -19,6 +19,7 @@ import org.corfudb.infrastructure.logreplication.replication.send.logreader.Defa
 import org.corfudb.infrastructure.logreplication.replication.send.logreader.ReadProcessor;
 import org.corfudb.infrastructure.logreplication.runtime.LogReplicationClient;
 import org.corfudb.infrastructure.logreplication.utils.LogReplicationConfigManager;
+import org.corfudb.infrastructure.logreplication.utils.LogReplicationUpgradeManager;
 import org.corfudb.runtime.CorfuRuntime;
 import org.corfudb.runtime.CorfuRuntime.CorfuRuntimeParameters;
 
@@ -70,15 +71,18 @@ public class LogReplicationSourceManager {
      * @param metadataManager Replication Metadata Manager
      */
     public LogReplicationSourceManager(LogReplicationRuntimeParameters params, LogReplicationClient client,
-        LogReplicationMetadataManager metadataManager, LogReplicationConfigManager tableManagerPlugin,
-        ReplicationSession replicationSession) {
-        this(params, metadataManager, new CorfuDataSender(client), tableManagerPlugin, replicationSession);
+                                       LogReplicationMetadataManager metadataManager,
+                                       LogReplicationConfigManager configManager,
+                                       LogReplicationUpgradeManager upgradeManager,
+                                       ReplicationSession replicationSession) {
+        this(params, metadataManager, new CorfuDataSender(client), configManager, upgradeManager, replicationSession);
     }
 
     @VisibleForTesting
     public LogReplicationSourceManager(LogReplicationRuntimeParameters params,
-        LogReplicationMetadataManager metadataManager, DataSender dataSender,
-        LogReplicationConfigManager tableManagerPlugin, ReplicationSession replicationSession) {
+                                       LogReplicationMetadataManager metadataManager, DataSender dataSender,
+                                       LogReplicationConfigManager configManager, LogReplicationUpgradeManager upgradeManager,
+                                       ReplicationSession replicationSession) {
 
         // This runtime is used exclusively for the snapshot and log entry reader which do not require a cache
         // as these are one time operations.
@@ -113,10 +117,10 @@ public class LogReplicationSourceManager {
         this.metadataManager = metadataManager;
 
         // Ack Reader for Snapshot and LogEntry Sync
-        this.ackReader = new LogReplicationAckReader(this.metadataManager, tableManagerPlugin, runtime, replicationSession);
+        this.ackReader = new LogReplicationAckReader(this.metadataManager, configManager, runtime, replicationSession);
 
-        this.logReplicationFSM = new LogReplicationFSM(this.runtime, tableManagerPlugin, dataSender, readProcessor,
-            logReplicationFSMWorkers, ackReader, tableManagerPlugin, replicationSession);
+        this.logReplicationFSM = new LogReplicationFSM(this.runtime, configManager, dataSender, readProcessor,
+            logReplicationFSMWorkers, ackReader, upgradeManager, replicationSession);
 
         this.logReplicationFSM.setTopologyConfigId(params.getTopologyConfigId());
         this.ackReader.setLogEntryReader(this.logReplicationFSM.getLogEntryReader());
