+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/replication/fsm/WaitSnapshotApplyState.java
@@ -7,7 +7,7 @@ import org.corfudb.common.metrics.micrometer.MeterRegistryProvider;
 import org.corfudb.infrastructure.logreplication.DataSender;
 import org.corfudb.infrastructure.logreplication.replication.send.LogReplicationEventMetadata;
 import org.corfudb.infrastructure.logreplication.runtime.CorfuLogReplicationRuntime;
-import org.corfudb.infrastructure.logreplication.utils.LogReplicationConfigManager;
+import org.corfudb.infrastructure.logreplication.utils.LogReplicationUpgradeManager;
 import org.corfudb.runtime.LogReplication.LogReplicationMetadataResponseMsg;
 
 import java.util.Optional;
@@ -56,7 +56,7 @@ public class WaitSnapshotApplyState implements LogReplicationState {
     /**
      * Used for checking LR is in upgrading path or not
      */
-    private final LogReplicationConfigManager tableManagerPlugin;
+    private final LogReplicationUpgradeManager upgradeManager;
 
     /**
      * Base Snapshot Timestamp for current Snapshot Sync
@@ -74,14 +74,15 @@ public class WaitSnapshotApplyState implements LogReplicationState {
      *
      * @param logReplicationFSM log replication state machine
      */
-    public WaitSnapshotApplyState(LogReplicationFSM logReplicationFSM, DataSender dataSender, LogReplicationConfigManager tableManagerPlugin) {
+    public WaitSnapshotApplyState(LogReplicationFSM logReplicationFSM, DataSender dataSender,
+                                  LogReplicationUpgradeManager upgradeManager) {
         this.fsm = logReplicationFSM;
         this.dataSender = dataSender;
         this.snapshotSyncApplyMonitorExecutor = Executors.newSingleThreadScheduledExecutor(new ThreadFactoryBuilder()
                 .setDaemon(true)
                 .setNameFormat("snapshotSyncApplyVerificationScheduler")
                 .build());
-        this.tableManagerPlugin = tableManagerPlugin;
+        this.upgradeManager = upgradeManager;
     }
 
     @Override
@@ -123,12 +124,12 @@ public class WaitSnapshotApplyState implements LogReplicationState {
                     logEntrySyncState.setTransitionEventId(event.getEventId());
                     fsm.setBaseSnapshot(event.getMetadata().getLastTransferredBaseSnapshot());
                     fsm.setAckedTimestamp(event.getMetadata().getLastLogEntrySyncedTimestamp());
-                    if (tableManagerPlugin.isUpgraded()) {
+                    if (upgradeManager.isUpgraded()) {
                         // If LR is in upgrading path, it means this cycle of snapshot sync was triggered
                         // forcibly because LR detected a version mismatch. Flipping the flag back to false
                         // here to indicate that the upgrade path is completed.
                         log.info("Forced snapshot sync due to LR upgrade is COMPLETE.");
-                        tableManagerPlugin.resetUpgradeFlag();
+                        upgradeManager.resetUpgradeFlag();
                     }
                     log.info("Snapshot Sync apply completed, syncRequestId={}, baseSnapshot={}. Transition to LOG_ENTRY_SYNC",
                             event.getEventId(), event.getMetadata().getLastTransferredBaseSnapshot());
