+++ infrastructure/src/main/java/org/corfudb/infrastructure/logreplication/runtime/fsm/NegotiatingState.java
@@ -7,7 +7,7 @@ import org.corfudb.infrastructure.logreplication.replication.receive.LogReplicat
 import org.corfudb.infrastructure.logreplication.replication.send.LogReplicationEventMetadata;
 import org.corfudb.infrastructure.logreplication.runtime.CorfuLogReplicationRuntime;
 import org.corfudb.infrastructure.logreplication.runtime.LogReplicationClientRouter;
-import org.corfudb.infrastructure.logreplication.utils.LogReplicationConfigManager;
+import org.corfudb.infrastructure.logreplication.utils.LogReplicationUpgradeManager;
 import org.corfudb.runtime.LogReplication;
 import org.corfudb.runtime.LogReplication.LogReplicationMetadataResponseMsg;
 import org.corfudb.runtime.proto.service.CorfuMessage;
@@ -39,15 +39,15 @@ public class NegotiatingState implements LogReplicationRuntimeState {
 
     private final LogReplicationMetadataManager metadataManager;
 
-    private final LogReplicationConfigManager tableManagerPlugin;
+    private final LogReplicationUpgradeManager upgradeManager;
 
     public NegotiatingState(CorfuLogReplicationRuntime fsm, ThreadPoolExecutor worker, LogReplicationClientRouter router,
-                            LogReplicationMetadataManager metadataManager, LogReplicationConfigManager tableManagerPlugin) {
+                            LogReplicationMetadataManager metadataManager, LogReplicationUpgradeManager upgradeManager) {
         this.fsm = fsm;
         this.metadataManager = metadataManager;
         this.worker = worker;
         this.router = router;
-        this.tableManagerPlugin = tableManagerPlugin;
+        this.upgradeManager = upgradeManager;
     }
 
     @Override
@@ -77,7 +77,7 @@ public class NegotiatingState implements LogReplicationRuntimeState {
                 return null;
             case NEGOTIATION_COMPLETE:
                 log.info("Negotiation complete, result={}", event.getNegotiationResult());
-                if (tableManagerPlugin.isUpgraded()) {
+                if (upgradeManager.isUpgraded()) {
                     // Force a snapshot sync if an upgrade has been identified. This will guarantee that
                     // changes in the streams to replicate are captured by the destination.
                     log.info("A forced snapshot sync will be done as Source side LR has been upgraded.");
